Assignment2/Application/fastapi/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
openai==0.28.0
pandas==1.3.3
python-dotenv
protobuf==3.20.1
numpy==1.21.2
boto3
PyPDF2
altair==4.1.0
sqlalchemy
python-jose
bcrypt
passlib
psycopg2
python-multipart